Improve problem80 code style and implementation fixing the package configuration, renaming some variables, using a different javadoc and changing some methods modifiers from public static to private
