+++ src/main/java/com/github/pedrovgs/problem80/QuickSort.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.pedrovgs.problem79;
+package com.github.pedrovgs.problem80;
 
 import com.github.pedrovgs.sortingalgorithm.SortingAlgorithm;
 
@@ -23,53 +23,53 @@ import com.github.pedrovgs.sortingalgorithm.SortingAlgorithm;
  *
  * @author jsroyal
  */
-
-/**
- * Quick sort takes advantage for avarage case which O(nlogn);
- * It starts by comparing every two and so on.Final got sort list of the algorithms described here,
- * this is the first that
- * scales well to very large lists, because its worst-case running time is O(n^2). It is also
- * easily applied to lists, not only arrays, as it only requires sequential access, not random
- * access. However, it has additional O(n) space complexity, and involves a large number of
- * copies in simple implementations.
- */
-
 public class QuickSort extends SortingAlgorithm {
 
-  private int[] numbers;
-  private int number;
-
+  /**
+   * Quicksort (sometimes called partition-exchange sort) is an efficient sorting algorithm, serving
+   * as a systematic method for placing the elements of an array in order. Developed by Tony Hoare
+   * in 1959 and published in 1961, it is still a commonly used algorithm for sorting. When
+   * implemented well, it can be about two or three times faster than its main competitors, merge
+   * sort and heapsort.
+   *
+   * Quicksort is a comparison sort, meaning that it can sort items of any type for which a
+   * "less-than" relation (formally, a total order) is defined. In efficient implementations it is
+   * not a stable sort, meaning that the relative order of equal sort items is not preserved.
+   * Quicksort can operate in-place on an array, requiring small additional amounts of memory to
+   * perform the sorting.
+   *
+   * Mathematical analysis of quicksort shows that, on average, the algorithm takes O(n log n)
+   * comparisons to sort n items. In the worst case, it makes O(n2) comparisons, though this
+   * behavior is rare.
+   */
   @Override public void sort(int[] numbers) {
     validateInput(numbers);
-
-    this.numbers = numbers;
-    number = numbers.length;
-    quickSort(numbers, 0, number - 1);
+    quickSort(numbers, 0, numbers.length - 1);
   }
 
-  private static int partition(int[] a, int l, int r) {
-    int pivot = a[r];
-    while (l < r) {
-      while (a[l] < pivot) { //from left move
-        l++;
-      }
-      while (a[r] > pivot) {  //from right move
-        r--;
-      }
-      if (l <= r) {
-        int temp = a[l];
-        a[l] = a[r];
-        a[r] = temp;
-      }
-    }
-    return l; //pivot index
-  }
-
-  public static void quickSort(int[] a, int left, int right) {
+  private void quickSort(int[] numbers, int left, int right) {
     if (left < right) {
-      int pi = partition(a, left, right);  //pi index of pivot
-      quickSort(a, left, pi - 1);  //sort left of pivot
-      quickSort(a, pi, right);  //sort right of pivot
+      int pivotIndex = partition(numbers, left, right);
+      quickSort(numbers, left, pivotIndex - 1);  //sort left of pivot
+      quickSort(numbers, pivotIndex, right);  //sort right of pivot
     }
+  }
+
+  private int partition(int[] numbers, int left, int right) {
+    int pivot = numbers[right];
+    while (left < right) {
+      while (numbers[left] < pivot) {
+        left++;
+      }
+      while (numbers[right] > pivot) {
+        right--;
+      }
+      if (left <= right) {
+        int temp = numbers[left];
+        numbers[left] = numbers[right];
+        numbers[right] = temp;
+      }
+    }
+    return left; //pivot index
   }
 }
