+++ src/test/java/com/github/pedrovgs/problem80/QuickSortTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.pedrovgs.problem79;
+package com.github.pedrovgs.problem80;
 
+import com.github.pedrovgs.problem79.QuickSort;
 import com.github.pedrovgs.sortingalgorithm.SortingAlgorithm;
 import com.github.pedrovgs.sortingalgorithms.SortingAlgorithmTest;
